Split models file
