+++ goodpoop/api/models/restroom.py
@@ -1,0 +1,17 @@
+from django.db import models
+from goodpoop.api.models import Account
+
+
+class Restroom(models.Model):
+    name = models.CharField(max_length=255)
+    latitude = models.FloatField()
+    longitude = models.FloatField()
+    created_by = models.ForeignKey(
+        Account,
+        on_delete=models.SET_NULL,
+        blank=True,
+        null=True,
+        related_name="created_restrooms",
+    )
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
