+++ goodpoop/api/models/review.py
@@ -1,39 +1,6 @@
 from django.core.validators import MaxValueValidator, MinValueValidator
 from django.db import models
-
-
-class Account(models.Model):
-    email = models.EmailField(max_length=254, unique=True)
-    password = models.CharField(max_length=255)
-    name = models.CharField(max_length=255)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-
-class Restroom(models.Model):
-    name = models.CharField(max_length=255)
-    latitude = models.FloatField()
-    longitude = models.FloatField()
-    created_by = models.ForeignKey(
-        Account,
-        on_delete=models.SET_NULL,
-        blank=True,
-        null=True,
-        related_name="created_restrooms",
-    )
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-
-class CheckIn(models.Model):
-    restroom = models.ForeignKey(
-        Restroom, on_delete=models.CASCADE, related_name="check_ins"
-    )
-    account = models.ForeignKey(
-        Account, on_delete=models.CASCADE, related_name="check_ins"
-    )
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
+from goodpoop.api.models import Account, Restroom
 
 
 class Review(models.Model):
